+++ cmd/ttn-lw-cli/commands/end_devices.go
@@ -50,6 +50,8 @@ var (
 	getDefaultMACSettingsFlags = util.NormalizedFlagSet()
 	allEndDeviceSetFlags       = util.NormalizedFlagSet()
 	allEndDeviceSelectFlags    = util.NormalizedFlagSet()
+	listBandsFlags             = util.NormalizedFlagSet()
+	listPhyVersionFlags        = util.NormalizedFlagSet()
 	getNetIDFlags              = util.NormalizedFlagSet()
 	getDevAddrPrefixesFlags    = util.NormalizedFlagSet()
 
@@ -1417,6 +1419,57 @@ This command may take end device identifiers from stdin.`,
 			return io.Write(os.Stdout, config.OutputFormat, res)
 		},
 	}
+	endDevicesListBandsCommand = &cobra.Command{
+		Use:               "list-bands",
+		Short:             "List available band definitions",
+		PersistentPreRunE: preRun(),
+		RunE: func(cmd *cobra.Command, _ []string) error {
+			if !config.NetworkServerEnabled {
+				return errNetworkServerDisabled.New()
+			}
+
+			req := &ttnpb.ListBandsRequest{}
+			_, err := req.SetFromFlags(cmd.Flags(), "")
+			if err != nil {
+				return err
+			}
+			ns, err := api.Dial(ctx, config.NetworkServerGRPCAddress)
+			if err != nil {
+				return err
+			}
+			res, err := ttnpb.NewConfigurationClient(ns).ListBands(ctx, req)
+			if err != nil {
+				return err
+			}
+			return io.Write(os.Stdout, config.OutputFormat, res)
+		},
+	}
+	endDevicesListPhyVersionsCommand = &cobra.Command{
+		Use:               "list-phy-versions",
+		Aliases:           []string{"get-phy-versions"},
+		Short:             "List supported phy versions",
+		PersistentPreRunE: preRun(),
+		RunE: func(cmd *cobra.Command, _ []string) error {
+			if !config.NetworkServerEnabled {
+				return errNetworkServerDisabled.New()
+			}
+
+			req := &ttnpb.GetPhyVersionsRequest{}
+			_, err := req.SetFromFlags(cmd.Flags(), "")
+			if err != nil {
+				return err
+			}
+			ns, err := api.Dial(ctx, config.NetworkServerGRPCAddress)
+			if err != nil {
+				return err
+			}
+			res, err := ttnpb.NewConfigurationClient(ns).GetPhyVersions(ctx, req)
+			if err != nil {
+				return err
+			}
+			return io.Write(os.Stdout, config.OutputFormat, res)
+		},
+	}
 )
 
 func init() {
@@ -1424,6 +1477,8 @@ func init() {
 	ttnpb.AddSetFlagsForGetDefaultMACSettingsRequest(getDefaultMACSettingsFlags, "", false)
 	ttnpb.AddSelectFlagsForEndDevice(allEndDeviceSelectFlags, "", false)
 	ttnpb.AddSetFlagsForEndDevice(allEndDeviceSetFlags, "", false)
+	ttnpb.AddSetFlagsForListBandsRequest(listBandsFlags, "", false)
+	ttnpb.AddSetFlagsForGetPhyVersionsRequest(listPhyVersionFlags, "", false)
 
 	allEndDeviceSelectFlags.VisitAll(func(flag *pflag.Flag) {
 		fieldName := toUnderscore.Replace(flag.Name)
@@ -1537,6 +1592,12 @@ func init() {
 
 	endDevicesGetDevAddrPrefixesCommand.Flags().AddFlagSet(getDevAddrPrefixesFlags)
 	endDevicesCommand.AddCommand(endDevicesGetDevAddrPrefixesCommand)
+
+	endDevicesListBandsCommand.Flags().AddFlagSet(listBandsFlags)
+	endDevicesCommand.AddCommand(endDevicesListBandsCommand)
+
+	endDevicesListPhyVersionsCommand.Flags().AddFlagSet(listPhyVersionFlags)
+	endDevicesCommand.AddCommand(endDevicesListPhyVersionsCommand)
 
 	Root.AddCommand(endDevicesCommand)